@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.20)
 project(SpaceInvaders)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
 set(SOURCE_FILES main.cpp View/Menu.cpp View/Menu.h View/Info.cpp View/Info.h Controller/MenuController.cpp Controller/MenuController.h)
 
 add_executable(SpaceInvaders ${SOURCE_FILES})
