@@ -1,6 +1,5 @@
 #ifndef SPACEINVADERS_MENU_H
 #define SPACEINVADERS_MENU_H
-#include "../Controller/MenuController.h"
 #include <iostream>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
@@ -13,15 +12,28 @@ class Menu{
 public:
     static const int WIDTH = 1280;
     static const int HEIGHT = 1400;
-
+    //Constructor & Destructor
     Menu();
+    virtual ~Menu();
+
+    //functions
+    void pollEvents();
+    void update();
+    void render();
     void run();
     static void centerText(Text& text, float height);
     static void centerSprite(Sprite& sprite, float height);
 private:
-    int record;
-    int i;
+    //Private functions
+    void initVariables();
+    void initWindow();
+    const bool running() const;
+
+    //Variables
+    unique_ptr<RenderWindow> window;
+    VideoMode videoMode;
     Event event;
+
     Font f1;
     Text title;
     Text text;
@@ -32,7 +44,8 @@ private:
     Sprite sprShip;
     SoundBuffer buffer;
     Sound sound;
-    unique_ptr<MenuController> mc;
+    int record;
+    int i;
 };
 
 #endif //SPACEINVADERS_MENU_H
